refactor: Palette 패키지 분리
